% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje.R
\name{clean_sales_ts}
\alias{clean_sales_ts}
\title{Wyczysc dane sprzedazowe}
\usage{
clean_sales_ts(data)
}
\arguments{
\item{data}{dane sprzedazowe}
}
\value{
wyczyszczona tabela
}
\description{
Funkcja laczy duplikaty przez sumowanie sprzedazy, uzupelnia brakujace dni wartoscia 0 i sortuje dane.
}
