% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje.R
\name{compute_sales_metrics}
\alias{compute_sales_metrics}
\title{Oblicz metryki biznesowe}
\usage{
compute_sales_metrics(data, window = 7)
}
\arguments{
\item{data}{wyczyszczone dane}

\item{window}{liczba dni do sredniej kroczacej}
}
\value{
lista z tabela metryk i szeregiem czasowym
}
\description{
Funkcja liczy sprzedaz calkowita, srednia, zmiennosc, udzial promocji, srednia kroczaca, szczyty i zmiane procentowa.
}
