% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje.R
\name{create_management_summary}
\alias{create_management_summary}
\title{Utworz podsumowanie dla menedzera}
\usage{
create_management_summary(data)
}
\arguments{
\item{data}{dane sprzedazowe}
}
\value{
lista z rankingami i tekstem podsumowania
}
\description{
Funkcja tworzy ranking sklepow, ranking kategorii, kategorie rosnace i spadajace oraz ocene promocji.
}
