% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje.R
\name{create_prognosis}
\alias{create_prognosis}
\title{Stworz prognoze sprzedazy ARIMA i Prophet}
\usage{
create_prognosis(
  data,
  store,
  category,
  horizon = 30,
  use_prophet = TRUE,
  history_days = 120
)
}
\arguments{
\item{data}{dane sprzedazowe}

\item{store}{numer sklepu}

\item{category}{kategoria produktu}

\item{horizon}{liczba dni prognozy}

\item{use_prophet}{czy uruchamiac model Prophet}

\item{history_days}{liczba ostatnich dni historii na wykresie}
}
\value{
lista z danymi, modelami, prognozami, porownaniem i wykresem
}
\description{
Funkcja prognozuje sprzedaz dla wybranego sklepu i kategorii. Domyslnie wykorzystuje ARIMA oraz Prophet. W raporcie PDF mozna uzyc use_prophet = FALSE, aby raport tworzyl sie szybciej.
}
