% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje.R
\name{load_sales_data}
\alias{load_sales_data}
\title{Wczytaj dane sprzedazowe}
\usage{
load_sales_data(
  train_path = "data/train.csv",
  stores_path = "data/stores.csv",
  holidays_path = "data/holidays_events.csv"
)
}
\arguments{
\item{train_path}{sciezka do pliku train.csv}

\item{stores_path}{sciezka do pliku stores.csv}

\item{holidays_path}{sciezka do pliku holidays_events.csv}
}
\value{
tabela z danymi sprzedazowymi
}
\description{
Funkcja wczytuje trzy pliki CSV: dane sprzedazy, dane sklepow oraz dane o swietach. Nastepnie laczy je w jedna tabele analityczna.
}
