% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje.R
\name{plot_sales_trends}
\alias{plot_sales_trends}
\title{Narysuj trend sprzedazy}
\usage{
plot_sales_trends(data, store = NULL, category = NULL)
}
\arguments{
\item{data}{dane sprzedazowe}

\item{store}{wybrany numer sklepu}

\item{category}{wybrana kategoria albo kategorie}
}
\value{
wykres ggplot
}
\description{
Funkcja tworzy wykres trendu sprzedazy z 14-dniowa srednia kroczaca.
}
