% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje.R
\name{validate_sales_ts}
\alias{validate_sales_ts}
\title{Sprawdz jakosc danych}
\usage{
validate_sales_ts(data)
}
\arguments{
\item{data}{dane sprzedazowe}
}
\value{
lista z wynikami walidacji
}
\description{
Funkcja sprawdza braki, duplikaty, bledne daty, ujemna sprzedaz i brakujace dni w szeregach czasowych.
}
